requirements.txt
langchain==0.0.352
langchain-community==0.0.7
langchain-openai==0.0.2
chromadb==0.4.22
huggingface_hub==0.20.1
python-dotenv==1.0.0